refactor: repository의 필요없는 부분들을 삭제하고 메소드를 옮긴다.
