+++ src/main/java/com/dnd/accompany/domain/accompany/infrastructure/querydsl/AccompanyBoardRepositoryImpl.java
@@ -91,12 +91,4 @@ public class AccompanyBoardRepositoryImpl implements AccompanyBoardRepositoryCus
 
 		return fetchCount != null;
 	}
-
-	@Override
-	public Optional<String> findNickname(Long userId) {
-		return Optional.of(queryFactory
-			.select(user.nickname)
-			.from(user)
-			.fetchOne());
-	}
 }
