+++ src/main/java/com/dnd/accompany/domain/accompany/infrastructure/querydsl/AccompanyUserRepositoryImpl.java
@@ -41,4 +41,13 @@ public class AccompanyUserRepositoryImpl implements AccompanyUserRepositoryCusto
 			)
 			.fetchOne());
 	}
+
+	@Override
+	public Optional<String> findNickname(Long userId) {
+		return Optional.of(queryFactory
+			.select(user.nickname)
+			.from(user)
+			.where(user.id.eq(userId))
+			.fetchOne());
+	}
 }
