+++ src/main/java/com/dnd/accompany/domain/accompany/infrastructure/querydsl/interfaces/AccompanyBoardRepositoryCustom.java
@@ -14,6 +14,4 @@ public interface AccompanyBoardRepositoryCustom {
 	Slice<FindBoardThumbnailsResult> findBoardThumbnails(Pageable pageable, Region region);
 
 	boolean isHostOfBoard(Long userId, Long boardId);
-
-	Optional<String> findNickname(Long userId);
 }
